feat!: add options pattern to OSCAL catalog generation
This updates the Layer 1 and Layer 2 implementations of OSCAL
generation for a more cohesive approach. The API now uses an options
pattern instead of positional parameters, providing more flexibility
and consistency across OSCAL generation functions.

Breaking change: FromCatalog renamed to CatalogFromCatalog and now
uses WithControlHref option instead of positional controlHREF parameter.

Assisted by: Composer(Cursor AI)
